implement generalized userid and timestamps
This commit implements pattern matching for general user ids but its not
working perfectly. It also adds an optional flag to anonymize timestamps
which seems to work pretty well. For the user ids, the next step is to
anonymize them according to the specific type of log file we are working
with.

Signed-off-by: Vlad Arama <vlad.arama@ericsson.com>