split anonymizing of userids
This commit splits the anonymizing of the user ids according to their
specific patterns. For example in http and gc logs, the user ids start
with "- " and end with whitespace, which makes them easily detectable.

Ip addresses are well detected and anonymized, endpoints are still a bit finicky.
Function paths (`java.org.file.blabla...`) are not anonymized yet.

Signed-off-by: Vlad Arama <vlad.arama@ericsson.com>